refactoring
separeted tests and framework
created a main (unit_tests.c)
