+++ tests/00_launcher.c
@@ -6,22 +6,19 @@
 /*   By: mmillhof <mmillhof@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/06 11:32:13 by mmillhof          #+#    #+#             */
-/*   Updated: 2025/09/06 13:22:57 by mmillhof         ###   ########.fr       */
+/*   Updated: 2025/09/06 15:32:07 by mmillhof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "basic_tests.h"
 #include "libunit.h"
 
-int		launch_tests(t_unit_test *testlist);
-void	load_test(t_unit_test *testlist, char *name, void *test);
-
 int	strlen_launcher(void)
 {
 	t_unit_test	*testlist;
 	int			ret;
 
-	testlist = malloc(100 * sizeof(t_unit_test));
+	testlist = malloc(5 * sizeof(t_unit_test));
 	testlist[0].test = NULL;
 	load_test(testlist, "Basic test", &basic_test);
 	load_test(testlist, "NULL test", &null_test);
@@ -30,42 +27,4 @@ int	strlen_launcher(void)
 	ret = launch_tests(testlist);
 	free(testlist);
 	return (ret);
-}
-
-int	launch_tests(t_unit_test *testlist)
-{
-	int	i;
-	int	pid;
-	int	ret;
-	int	status;
-
-	i = 0;
-	while (testlist[i].test)
-	{
-		pid = fork();
-		if (pid == 0)
-		{
-			ret = testlist[i].test();
-			exit (ret);
-		}
-		printf("testing: %s\n", testlist[i].name);
-		while (wait(&status) > 0)
-			;
-		printf("status: %i:\n\n", WEXITSTATUS(status));
-		i++;
-	}
-	return (0);
-}
-
-void	load_test(t_unit_test *testlist, char *name, void *test)
-{
-	int	i;
-
-	i = 0;
-	while (testlist[i].test)
-		i++;
-	testlist[i].name = name;
-	testlist[i].test = test;
-	testlist[i + 1].test = NULL;
-	return ;
 }
