+++ tests/02_null_test.c
