+++ tests/includes/basic_tests.h
@@ -6,15 +6,17 @@
 /*   By: mmillhof <mmillhof@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/06 12:10:04 by mmillhof          #+#    #+#             */
-/*   Updated: 2025/09/06 12:35:55 by mmillhof         ###   ########.fr       */
+/*   Updated: 2025/09/06 15:53:45 by mmillhof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef BASIC_TESTS_H
 # define BASIC_TESTS_H
 
+# include "libunit.h"
 # include <string.h>
 
+int	strlen_launcher(void);
 int	basic_test(void);
 int	null_test(void);
