+++ unit_test.c
@@ -1,32 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libunit.h                                          :+:      :+:    :+:   */
+/*   unit_test.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mmillhof <mmillhof@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/09/06 11:30:07 by mmillhof          #+#    #+#             */
-/*   Updated: 2025/09/06 15:42:50 by mmillhof         ###   ########.fr       */
+/*   Created: 2025/09/06 15:34:29 by mmillhof          #+#    #+#             */
+/*   Updated: 2025/09/06 15:36:24 by mmillhof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LIBUNIT_H 
-# define LIBUNIT_H
+#include "basic_tests.h"
 
-# include <sys/types.h>
-# include <sys/wait.h>
-# include <unistd.h>
-# include <stdlib.h>
-# include <stdio.h>
-
-typedef struct s_unit_test
+int	main(int argc, char **argv)
 {
-	int		(*test)(void);
-	char	*name;
-}	t_unit_test;
+	(void) argc;
+	(void) argv;
+	return (strlen_launcher());
+}
 
-int		strlen_launcher(void);
-int		launch_tests(t_unit_test *testlist);
-void	load_test(t_unit_test *testlist, char *name, void *test);
-
-#endif
